+++ CHANGELOG.md
@@ -107,6 +107,7 @@
 - Suppression de la possibilité de mettre la règle 'role_id_e: no-role' sur les connecteurs. Il faut remplacer par la règle 'automatique: true' pour obtenir le même comportement
 - Suppression des élements relatifs au centre de gestions sur la version standard. Il est possible de les remettre avec la feature flag CDGFeature.
 - Suppression de l'URL dans l'onglet de retour du SAE
+- Suppression du support MySQL. La seule base supportée est MariaDB.
 
 # [3.1.17] - 2022-09-12